fix and organize tests
