+++ server/package.json
@@ -27,7 +27,7 @@
     "yaml": "^2.2.2"
   },
   "scripts": {
-    "tests": "mocha --require ts-node/register test/**/*.ts"
+    "tests": "mocha --require ts-node/register test/**/*.test.ts"
   },
   "devDependencies": {
     "@types/chai": "^4.3.0",
@@ -37,7 +37,8 @@
     "@types/yaml": "1.9.7",
     "chai": "^4.3.6",
     "mocha": "^9.2.1",
-    "ts-node": "^10.5.0",
+    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
     "typescript": "^4.7.4"
   }
 }
